requirements.txt
argon2-cffi
miscreant
msgpack
pip-chill
pre-commit
python-gnupg
